refactor: npm 배포를 위해 lib 폴더로 이동 및 index 파일에 경로 추가
