+++ components/src/lib/assets/close.png
