+++ components/src/lib/index.ts
@@ -1,1 +1,4 @@
-export { default as Modal } from "./modal/Modal";
+export { default as Modal } from './modal/Modal';
+export { default as AlertModal } from './modal/AlertModal';
+export { default as ComfirmModal } from './modal/ConfirmModal';
+export { default as PromptModal } from './modal/PromptModal';
